Adding ratings, ratings tests, and authentication tests.
The chunk of authentication/authorization tests at the end of brew.spec could probably be rewritten so that they read a bit more like BDD tests.  I was experimenting with exactly how the before(), describe(), and it() calls behaved when nested in strange ways.

It took awhile, but I finally figured out how to test using credentials, setting the authorization header.